Add files via upload
INTRODUCTION

I examined and contrasted the rate of assault crime in Ontario, Quebec, and British Columbia during the last ten years in this study to understand its trends and patterns in these provinces. I made use of Microsoft Power BI to generate visualizations that made it simpler to compare and explore the data. These visualizations assisted in making comparing the crime rates between the provinces, identifying the top three cities in each province with the highest assault rates, evaluating the relationship between youth crime and adult crime in the provinces, and forecasting the assault crime rate for the following year in each province.



JUSTIFICATION FOR THE VISUALIZATION TOOL

With a user-friendly interface, interactive visualization, and drag-and-drop capabilities, Microsoft Power BI is a great data visualization tool for analysis. It offers an effective dashboard and report sharing and interacts with numerous data sources for effective analysis of crime statistics data across numerous provinces. We can effectively deal with the pertinent crime statistics data from many provinces thanks to its interaction with a variety of data sources.



EXPLANATION OF DATA ANALYSIS/VISUALIZATION APPROACH

To analyze the crime rates between Ontario, Quebec, and British Columbia, the data will be transformed and cleaned using power query to make it suitable for analysis. Once the data is ready, visualizations will be created using different chart types such as bar charts, scatterplot, and map. Then a comparison will be made using these visualizations to compare crime rates

between the provinces over the past 10 years and identify the top 3 cities with the highest crime rates in each province. Additionally, the correlation between youth crime and adult crime by cities would be examined. Lastly, predictive analytics techniques would be used to forecast the crime rate for the upcoming year in Ontario, Quebec, and British Columbia.



CONCLUSION

In conclusion, our analysis using Microsoft Power BI has provided valuable insights into the crime rates in Ontario, Quebec, and British Columbia. We have compared the crime rates between the provinces, identified the top 3 cities with the highest crime rates, examined the correlation between youth crime and adult crime in high crime rate cities, and predicted the crime rates for the upcoming year. Overall, from our forecast, there tends to be an increase in

the rate of crime for the next year compared to the previous years in Ontario, while there is a slight decrease in the crime rate for British Columbia and Quebec. Some strict laws should be enforced in addition to holding seminars in cities to help curb the rate of crime as well as the effect it would have on the These findings can help the RCMP in formulating targeted strategies to better support and work with the police in different provinces, ultimately contributing to a safer and more secure Canada.